added alerts

+++ view/admin/admin_crud/index.php
@@ -109,7 +109,7 @@ include ("../../../config.php");
                                                     class="btn btn-outline-warning m-1">Edit</a>
                                             </td>
                                             <td class="border-bottom-0">
-                                                <a href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
+                                                <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
                                                     class="btn btn-outline-danger m-1">Delete</a>
                                             </td>
 
@@ -174,7 +174,7 @@ include ("../../../config.php");
                                                         class="btn btn-outline-warning m-1">Edit</a>
                                                 </td>
                                                 <td class="border-bottom-0">
-                                                    <a href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
+                                                     <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?admin_id=<?php echo $row["admin_id"] ?>"
                                                         class="btn btn-outline-danger m-1">Delete</a>
                                                 </td>
