+++ view/admin/formateur_crud/index.php
@@ -123,7 +123,7 @@ include ("../../../config.php");
                                                     class="btn btn-outline-warning m-1">Edit</a>
                                             </td>
                                             <td class="border-bottom-0">
-                                                <a href="delete.php?formateur_id=<?php echo $row["formateur_id"] ?>"
+                                                 <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?formateur_id=<?php echo $row["formateur_id"] ?>"
                                                     class="btn btn-outline-danger m-1">Delete</a>
                                             </td>
 
@@ -195,7 +195,7 @@ include ("../../../config.php");
                                                         class="btn btn-outline-warning m-1">Edit</a>
                                                 </td>
                                                 <td class="border-bottom-0">
-                                                    <a href="delete.php?formateur_id=<?php echo $row["formateur_id"] ?>"
+                                                     <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?formateur_id=<?php echo $row["formateur_id"] ?>"
                                                         class="btn btn-outline-danger m-1">Delete</a>
                                                 </td>
