+++ view/admin/module_crud/index.php
@@ -124,7 +124,7 @@ include ("../../../config.php");
                                                     class="btn btn-outline-warning m-1">Edit</a>
                                             </td>
                                             <td class="border-bottom-0">
-                                                <a href="delete.php?module_id=<?php echo $row["module_id"] ?>"
+                                                 <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?module_id=<?php echo $row["module_id"] ?>"
                                                     class="btn btn-outline-danger m-1">Delete</a>
                                             </td>
 
@@ -185,7 +185,7 @@ include ("../../../config.php");
                                                         class="btn btn-outline-warning m-1">Edit</a>
                                                 </td>
                                                 <td class="border-bottom-0">
-                                                    <a href="delete.php?module_id=<?php echo $row["module_id"] ?>"
+                                                     <a onclick="return confirm('Are you sure you want to delete?')"  href="delete.php?module_id=<?php echo $row["module_id"] ?>"
                                                         class="btn btn-outline-danger m-1">Delete</a>
                                                 </td>
